jakeLi0311Code

+++ repo/jcart-administration-back/src/main/java/io/ljh/jcartadministrationback/po/OrderHistory.java
@@ -23,7 +23,7 @@ public class OrderHistory {
         this.orderHistoryId = orderHistoryId;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
