+++ repo/jcart-administration-back/src/main/java/io/ljh/jcartadministrationback/po/ReturnHistory.java
@@ -15,7 +15,7 @@ public class ReturnHistory {
 
     private Boolean customerNotified;
 
-    public Integer getReturnHistoryId() {
+    public Long getReturnHistoryId() {
         return returnHistoryId;
     }
